# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The Book-Store App",
        "version": "1.0.0",
        "title": "Book-Store-App"
    },
    "host": "localhost:8000",
...
    "schemes": [
        "http"
    ],
    "consumes":[
        "application/json"
    ],
    "produces":[
        "application/json"
    ],
    "paths": {
...
        "/books/{id}":{
            "put": {
                "tags": [
                    "Book Store App: Update A Perticular Book"
                ],
                "summary": "Update an existing Book",
                "description": "find by ID and update a Perticular Book",
                "operationId": "updatebooks",
                "parameters": [
                    {
                        "name": " Books ",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "author": {
                                    "type": "string"
                                },
                                "title": {
                                    "type": "string"
                                },
                                "image": {
                                    "type": "string"
                                },
                                "quantity": {
                                    "type": "string"
                                },
                                "price": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                } 
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book updated successfully"
                    },
                    "404": {
                        "description": "Failed To Update Book"
                    },
                    "500": {
                        "description": "could not find Book"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Book Store App: Delete A Perticular Book"
                ],
                "summary": "Delete a Book",
                "description": "Delete a Book by its Id",
                "operationId": "Delete Book",
                "responses": {
                    "200": {
                        "description": "Book deleted successfully"
                    },
                    "404": {
                        "description": "Failed To Delete Book"
                    },
                    "500": {
                        "description": "Could Not Find Book"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}